+++ crypto/CMakeLists.txt
@@ -323,9 +323,9 @@ if(ENABLE_DILITHIUM)
   set(
     DILITHIUM_SOURCES
 
-    dilithium/pqdsa.c
-    dilithium/p_pqdsa.c
-    dilithium/p_pqdsa_asn1.c
+    fipsmodule/pqdsa/pqdsa.c
+    fipsmodule/evp/p_pqdsa.c
+    evp_extra/p_pqdsa_asn1.c
     dilithium/ml_dsa.c
   )
 endif()
